Rendering basic triangle Added Basic Shapes precursor Moved glClearColor to be executed only once as that's how it's supposed to be Moved glClear before drawing

+++ Engine/Source/Renderer/OpenGL/Buffer/IndexBuffer.h
@@ -14,5 +14,5 @@ public:
 	void SetIndices(const uint32_t* inIndices, const int32_t inCount, const GLenum inBufferAccessType);
 
 public:
-	size_t IndicesCount;
+	int32_t IndicesCount;
 };
